+++ R_Net_Training_GIoU.py
@@ -60,9 +60,9 @@ if(not os.path.exists(f'weights/{weights_dir}')):
     print('[INFO] Created weight directory ...')
     os.mkdir(f'weights/{weights_dir}')
     
-if(os.path.exists(rnet_tensorboard_logdir)):
-    print('[INFO] Clearing R-Net log directory ... ')
-    shutil.rmtree(rnet_tensorboard_logdir)
+    if(os.path.exists(rnet_tensorboard_logdir)):
+        print('[INFO] Clearing R-Net log directory ... ')
+        shutil.rmtree(rnet_tensorboard_logdir)
 
 ### Loading dataset ###
 ### Creating the train loader ###
@@ -82,14 +82,21 @@ configs = {
     'input_shape' : input_dim*2,
     'batch_norm' : True,
     'dropout' : True,
-    'n_classes' : n_classes
+    'n_classes' : n_classes,
+    'initial_epoch' : 0
 }
 rnet = build_pnet_model(input_shape=configs['input_shape'], batch_norm=configs['batch_norm'], dropout=configs['dropout'],
                         n_classes=configs['n_classes'], l2_norm=True)
 
-print(f'[INFO] Storing R-Net configuration to {rnet_configs}')
-with open(rnet_configs, 'w') as config_file:
-    json.dump(configs, config_file, indent=4, sort_keys=True)
+if(not os.path.exists(rnet_configs)):
+    print(f'[INFO] Storing R-Net configuration to {rnet_configs}')
+    with open(rnet_configs, 'w') as config_file:
+        json.dump(configs, config_file, indent=4, sort_keys=True)
+else:
+    # Reload the configs
+    f = open(rnet_configs, 'r')
+    configs = json.load(f)
+    f.close()
 
 print(rnet.summary())
 
@@ -97,12 +104,21 @@ print(rnet.summary())
 steps_per_epoch = train_loader.dataset_len
 validation_steps = val_loader.dataset_len
 
-train(rnet, train_dataset, val_dataset, rnet_weights, 
-        logdir=rnet_tensorboard_logdir,
-        n_classes=n_classes, 
-        box_reg='mse',
-        steps_per_epoch=steps_per_epoch, 
-        validation_steps=validation_steps, 
-        epochs=epochs, 
-        make_conf_map=False)
-print('[INFO] Training halted, plotting training history ... ')
+while(True):
+    try:
+        last_epoch = train(rnet, train_dataset, val_dataset, rnet_weights, 
+                logdir=rnet_tensorboard_logdir,
+                n_classes=n_classes, 
+                box_reg='mse',
+                steps_per_epoch=steps_per_epoch, 
+                validation_steps=validation_steps, 
+                epochs=epochs, 
+                make_conf_map=False)
+        
+        # Save last epoch as initial epoch to configs
+        configs['initial_epoch'] = last_epoch
+        with open(rnet_configs, 'w') as config_file:
+            print('[INFO] Saving configs ...')
+            json.dump(configs, config_file, indent=4, sort_keys=True)
+    except KeyboardInterrupt:
+        break
